requirements.txt
certifi==2020.6.20
chardet==3.0.4
idna==2.9
requests==2.24.0
six==1.15.0
urllib3==1.25.9
websocket-client==0.57.0
